Add active class to the currently active tab content, as well as the tab itself
If this is not done, then the active class remains on the tab that was initially visible on page load - even if this tab is now hidden and no longer active. This is confusing when trying to target CSS as a the selector "ul.tabs-content > li.active" , may not be visible.

Also removed unused tabsContent variable.